requirements.txt
actionlib==1.14.0
angles==1.9.13
base_local_planner==1.17.3
bondpy==1.8.6
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
cloudpickle==2.2.1
contourpy==1.0.7
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cv-bridge==1.16.2
cycler==0.11.0
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.3
fonttools==4.39.4
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gym==0.26.2
gym-notices==0.0.8
image-geometry==1.16.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
interactive-markers==1.12.0
joblib==1.2.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.4
laser_geometry==1.6.7
matplotlib==3.7.1
message-filters==1.16.0
moveit-core==1.1.11
moveit-python==0.4.5
moveit-ros-planning-interface==1.1.11
moveit-ros-visualization==1.1.11
networkx==3.1
numpy==1.24.3
openni2_launch==1.6.0
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-qt-binding==0.4.4
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource_retriever==1.12.7
scikit-learn==1.2.2
scipy==1.10.1
sensor-msgs==1.13.1
six==1.16.0
sklearn==0.0.post5
smach==2.5.1
smach-ros==2.5.1
smclib==1.8.6
sound-play==0.3.16
srdfdom==0.6.3
threadpoolctl==3.1.0
topic-tools==1.16.0
urdfdom-py==0.4.6
xacro==1.14.15
zipp==3.15.0